working first test for fluid sdf

+++ common/inc/vtk_writer.h
@@ -20,6 +20,6 @@ namespace learnSPH
                                const std::vector<Eigen::Vector3d>& vector_data);
 
     void save_mesh_to_vtk(std::string path,
-                          const std::vector<Eigen::Vector3f>& vertices,
+                          const std::vector<Eigen::Vector3d>& vertices,
                           const std::vector<std::array<int, 3>>& triangles);
 }
